A change to how the generated Toggle Tag plugin handles mixed selections
Previously, if the first item had the tag, it would remove the tag from all selected items, and if not add the tag to the entire selection.

Now it adds the tag if *any* of the selected items are missing the tag, and removes otherwise.

(I'm going to update the standalone ToggleTag.omnijs plugin shortly to match this behavior)